Module Extend_zone.Set

module Elt : sig ... end

include Core_kernel.Set_intf.S with module Elt := Elt

module Tree : sig ... end

include Core_kernel.Set_intf.S_plain with module Elt := Elt and module Tree := Tree

type t = (Elt.t, Elt.comparator_witness) Base.Set.t

val compare : t -> t -> Base.Int.t

type named = (Elt.t, Elt.comparator_witness) Core_kernel.Set_intf.Named.t

include Core_kernel.Set_intf.Creators_and_accessors0 with type ('a, 'b) set := ('a, 'b) Base.Set.t with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Core_kernel.Set_intf.Accessors0 with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Core_kernel.Set_intf.Set.Accessors0 with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Base.Container.S0 with type t := t with type elt := Elt.t

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:(Elt.t -> unit) -> unit

iter must allow exceptions raised in f to escape, terminating the iteration cleanly. The same holds for all functions below taking an f.

val fold : t -> init:'accum -> f:('accum -> Elt.t -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : t -> init:'accum -> f:('accum -> Elt.t -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val exists : t -> f:(Elt.t -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : t -> f:(Elt.t -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : t -> f:(Elt.t -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> t -> f:(Elt.t -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : t -> f:(Elt.t -> bool) -> Elt.t option

Returns as an option the first element for which f evaluates to true.

val find_map : t -> f:(Elt.t -> 'a option) -> 'a option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : t -> Elt.t list

val to_array : t -> Elt.t array

val invariants : t -> bool

val mem : t -> Elt.t -> bool

val add : t -> Elt.t -> t

val remove : t -> Elt.t -> t

val union : t -> t -> t

val inter : t -> t -> t

val diff : t -> t -> t

val symmetric_diff : t -> t -> (Elt.t, Elt.t) Base.Either.t Base.Sequence.t

val compare_direct : t -> t -> int

val equal : t -> t -> bool

val is_subset : t -> of_:t -> bool

val are_disjoint : t -> t -> bool

module Named : sig ... end

val fold_until : t -> init:'b -> f:('b -> Elt.t -> ('b, 'final) Base__Set_intf.Continue_or_stop.t) -> finish:('b -> 'final) -> 'final

val fold_right : t -> init:'b -> f:(Elt.t -> 'b -> 'b) -> 'b

val iter2 : t -> t -> f:([ `Left of Elt.t | `Right of Elt.t | `Both of Elt.t * Elt.t ] -> unit) -> unit

val filter : t -> f:(Elt.t -> bool) -> t

val partition_tf : t -> f:(Elt.t -> bool) -> t * t

val elements : t -> Elt.t list

val min_elt : t -> Elt.t option

val min_elt_exn : t -> Elt.t

val max_elt : t -> Elt.t option

val max_elt_exn : t -> Elt.t

val choose : t -> Elt.t option

val choose_exn : t -> Elt.t

val split : t -> Elt.t -> t * Elt.t option * t

val group_by : t -> equiv:(Elt.t -> Elt.t -> bool) -> t list

val find_exn : t -> f:(Elt.t -> bool) -> Elt.t

val nth : t -> int -> Elt.t option

val remove_index : t -> int -> t

val to_tree : t -> Tree.t

val to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:Elt.t -> ?less_or_equal_to:Elt.t -> t -> Elt.t Base.Sequence.t

val binary_search : t -> compare:(Elt.t -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> Elt.t option

val binary_search_segmented : t -> segment_of:(Elt.t -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> Elt.t option

val merge_to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:Elt.t -> ?less_or_equal_to:Elt.t -> t -> t -> (Elt.t, Elt.t) Base.Sequence.Merge_with_duplicates_element.t Base.Sequence.t

val to_map : t -> f:(Elt.t -> 'data) -> (Elt.t, 'data, Elt.comparator_witness) Base.Map.t

val quickcheck_observer : Elt.t Core_kernel.Quickcheck.Observer.t -> t Core_kernel.Quickcheck.Observer.t

val quickcheck_shrinker : Elt.t Core_kernel.Quickcheck.Shrinker.t -> t Core_kernel.Quickcheck.Shrinker.t

include Core_kernel.Set_intf.Creators0 with type t := t with type tree := Tree.t with type elt := Elt.t with type comparator_witness := Elt.comparator_witness with type ('a, 'b) set := ('a, 'b) Base.Set.t

include Core_kernel.Set_intf.Set.Creators0 with type t := t with type tree := Tree.t with type elt := Elt.t with type comparator_witness := Elt.comparator_witness with type ('a, 'b) set := ('a, 'b) Base.Set.t

val empty : t

val singleton : Elt.t -> t

val union_list : t list -> t

val of_list : Elt.t list -> t

val of_array : Elt.t array -> t

val of_sorted_array : Elt.t array -> t Base.Or_error.t

val of_sorted_array_unchecked : Elt.t array -> t

val of_increasing_iterator_unchecked : len:int -> f:(int -> Elt.t) -> t

val stable_dedup_list : Elt.t list -> Elt.t list

val map : ('a, _) Base.Set.t -> f:('a -> Elt.t) -> t

val filter_map : ('a, _) Base.Set.t -> f:('a -> Elt.t option) -> t

val of_tree : Tree.t -> t

val of_hash_set : Elt.t Core_kernel.Hash_set.t -> t

val of_hashtbl_keys : (Elt.t, _) Core_kernel.Hashtbl.t -> t

val of_map_keys : (Elt.t, _, Elt.comparator_witness) Base.Map.t -> t

val quickcheck_generator : Elt.t Core_kernel.Quickcheck.Generator.t -> t Core_kernel.Quickcheck.Generator.t

module Provide_of_sexp (Elt : sig ... end) : sig ... end

module Provide_bin_io (Elt : sig ... end) : Core_kernel.Set_intf.Binable.S with type t := t

module Provide_hash (Elt : Base.Hasher.S with type t := Elt.t) : sig ... end

include Core_kernel.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core_kernel.Set_intf.Binable.S with type t := t

include Core_kernel.Set_intf.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t